finishing hammer nail tasks; updating TaskWatcher
The HammerNail scene now has two functioning tasks: placing the boards and then placing/hammering nails.  I opted to have placing and hammering the nails be a single task for simplicity.  An option has also been added to the TaskWatcher for updating the text string each frame.  This may not be desirable in all situations but is useful for updating a running count, which is needed in the HammerNails task.