disabling collider on main menu playervr rig
Opted to not replace this rig with a prefab because it would be necessary to get the laser pointers working on the prefabs (which may or may not introduce weird behavior in other scenes).  Instead the body colliders have just been disabled on the existing PlayerVR rig in the main menu.